+++ packages/modules/search/src/providers/search-medusa/services/medusa-search.ts
@@ -15,7 +15,7 @@ import {
   parseHighlights,
   toSearchDocument,
   toSearchFilter,
-  validateMedusaSearchOptions,
+  resolveMedusaSearchOptions,
   type FacetQuery,
   type IndexQuery,
   type IndexMultiQueryResponse,
@@ -53,11 +53,11 @@ export class MedusaSearchService extends AbstractSearchProviderService {
   ) {
     super()
 
-    validateMedusaSearchOptions(options)
+    const resolvedOptions = resolveMedusaSearchOptions(options)
 
     this.logger_ = logger
-    this.options_ = options
-    this.client_ = new MedusaSearchClient(options)
+    this.options_ = resolvedOptions
+    this.client_ = new MedusaSearchClient(resolvedOptions)
   }
 
   /**
@@ -96,7 +96,6 @@ export class MedusaSearchService extends AbstractSearchProviderService {
         name: index.physical_name,
         schema: plan.schema,
         distance_metric: plan.options.distance_metric,
-        sharding: plan.options.sharding,
       })
       return this.task(index.physical_name)
     }