+++ packages/plugins/loyalty/src/workflows/hooks/after-order-credit-lines-created.ts
@@ -1,12 +1,11 @@
-import { StepResponse } from "@medusajs/framework/workflows-sdk";
-import { createOrderCreditLinesWorkflow } from "@medusajs/medusa/core-flows";
-import { confirmCartCreditLinesWorkflow } from "../carts/workflows/confirm-cart-credit-lines";
-import { refundCreditLinesWorkflow } from "../orders/workflows/refund-credit-lines";
+import { StepResponse } from "@medusajs/framework/workflows-sdk"
+import { createOrderCreditLinesWorkflow } from "@medusajs/medusa/core-flows"
+import { refundCreditLinesWorkflow } from "../orders/workflows/refund-credit-lines"
 
-(createOrderCreditLinesWorkflow.hooks as any).creditLinesCreated(
+;(createOrderCreditLinesWorkflow.hooks as any).creditLinesCreated(
   async (data, stepContext) => {
-    const { container, ...sharedContext } = stepContext;
-    const { credit_lines: createdCreditLines } = data;
+    const { container, ...sharedContext } = stepContext
+    const { credit_lines: createdCreditLines } = data
 
     const transaction = await refundCreditLinesWorkflow.run({
       input: {
@@ -16,23 +15,23 @@ import { refundCreditLinesWorkflow } from "../orders/workflows/refund-credit-lin
       container,
       context: { ...sharedContext },
       throwOnError: true,
-    });
+    })
 
-    const { result } = transaction;
+    const { result } = transaction
 
-    return new StepResponse(result, stepContext.transactionId);
+    return new StepResponse(result, stepContext.transactionId)
   },
   async (transactionId, stepContext) => {
     if (!transactionId) {
-      return;
+      return
     }
 
-    const { container, ...sharedContext } = stepContext;
+    const { container, ...sharedContext } = stepContext
 
-    await confirmCartCreditLinesWorkflow(container).cancel({
+    await refundCreditLinesWorkflow(container).cancel({
       transactionId,
       container,
       context: { ...sharedContext },
-    });
+    })
   }
-);
+)